+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added support for `halfvec` and `sparsevec` types to Knex.js
 - Added `l1Distance`, `hammingDistance`, and `jaccardDistance` functions for Knex.js
 - Added `l1Distance`, `hammingDistance`, and `jaccardDistance` functions for Objection.js
+- Added `l1Distance`, `hammingDistance`, and `jaccardDistance` functions for Kysely
 - Dropped support for Node < 18
 
 ## 0.1.8 (2024-02-11)